Implemented changes as discussed in https://github.com/aeternity/dev-tools/pull/2
